@mitchehllh announced his intent to archive some of his unmaintained projects. So, migrate to the more actively developed version of mapstructure which is go-viper/mapstructure. This is required to update viper.

References:
* https://gist.github.com/mitchellh/90029601268e59a29e64e55bab1c5bdc
* https://github.com/mitchellh/mapstructure/issues/349

### Description

Perf:
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 571.83MiB/s  | 1.09MiB/s  |  79.6MiB/s   |  1.34MiB/s   |
|   PR   |  0.25MiB   | 585.93MiB/s  |  1.3MiB/s  |  82.31MiB/s  |  1.12MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3946.02MiB/s | 84.69MiB/s | 1525.52MiB/s |  88.21MiB/s  |
|   PR   | 48.828MiB  | 3971.01MiB/s | 87.67MiB/s | 1556.76MiB/s |  84.33MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4067.86MiB/s | 60.64MiB/s | 1284.29MiB/s |  35.8MiB/s   |
|   PR   | 976.562MiB | 3937.03MiB/s | 80.5MiB/s  | 1152.08MiB/s |  35.99MiB/s  |

### Link to the issue in case of a bug fix.
b/404140577

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - NA